Create Index of Available Months

Constructs a table for finding a given table of crime data or a set of tables (such as year to date or full year). This is largely needed for internal use when downloading tables, but is exported for reference and troubleshooting.

cs_create_index()

Value

A tibble with all available monthly crime tables, the iframe page they appear on, and their row number.

Examples

# create index
i <- cs_create_index()

# preview of index object
i

#> # A tibble: 148 x 6
#>     page row   value  year month date          
#>    <int> <chr> <chr> <dbl> <int> <chr>         
#>  1     1 02    1 02   2020     4 April 2020    
#>  2     1 03    1 03   2020     3 March 2020    
#>  3     1 04    1 04   2020     2 February 2020 
#>  4     1 05    1 05   2020     1 January 2020  
#>  5     1 06    1 06   2019    12 December 2019 
#>  6     1 07    1 07   2019    11 November 2019 
#>  7     1 08    1 08   2019    10 October 2019  
#>  8     1 09    1 09   2019     9 September 2019
#>  9     1 10    1 10   2019     8 August 2019   
#> 10     1 11    1 11   2019     7 July 2019     
#> # … with 138 more rows