Create Year List Object

cs_load_year is used to load a set of .csv files contained in the given directory. This should be used to load a full year worth of data or a partial year. There should be no more than 12 files in a given path, and all should correspond to the same year. All columns will be read in as character data in order to address inconsistencies in how the data are created. When cs_collapse is executed, variables will be converted numeric when doing so is applicable.

cs_load_year(path)

Arguments

path
A file path

Value

A year-list object containing 12 tibbles - one per month - worth of crime data stored within a list.

Examples

# create temporary directory
tmpdir <- tempdir()
fs::dir_create(paste0(tmpdir,"/data/"))

# load sample files into temporary directory
cs_example(path = paste0(tmpdir,"/data/"))

# prep sample files
cs_prep_year(path = paste0(tmpdir,"/data/"))

# load sample files
yearList17 <- cs_load_year(path = paste0(tmpdir,"/data/"))

# delete data
fs::dir_delete(paste0(tmpdir,"/data/"))

# print year-list object
yearList17

#> $April
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-019461 2017-04     01/01/200… Y          NA             NA              
#>  2 17-018445 2017-04     01/01/200… Y          NA             NA              
#>  3 17-014843 2017-04     01/01/201… Y          NA             NA              
#>  4 17-019313 2017-04     01/12/200… Y          NA             NA              
#>  5 17-019541 2017-04     01/12/201… Y          NA             NA              
#>  6 17-018072 2017-04     01/15/201… Y          NA             NA              
#>  7 17-017546 2017-04     01/15/201… Y          NA             NA              
#>  8 17-002915 2017-04     01/17/201… NA         NA             NA              
#>  9 17-017058 2017-04     01/19/201… Y          NA             NA              
#> 10 17-018971 2017-04     01/22/201… Y          NA             NA              
#> 11 17-020525 2017-04     01/24/201… Y          NA             NA              
#> 12 17-020524 2017-04     01/25/201… Y          NA             NA              
#> 13 17-020523 2017-04     01/28/201… Y          NA             NA              
#> 14 17-020522 2017-04     01/29/201… Y          NA             NA              
#> 15 17-004854 2017-04     01/29/201… Y          NA             NA              
#> 16 17-018542 2017-04     01/31/201… Y          NA             NA              
#> 17 17-016010 2017-04     02/01/201… Y          NA             NA              
#> 18 17-019585 2017-04     02/01/201… Y          NA             NA              
#> 19 17-016218 2017-04     02/02/201… Y          NA             NA              
#> 20 17-016895 2017-04     02/03/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $August
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-031380 2017-08     01/01/200… Y          NA             NA              
#>  2 17-042040 2017-08     01/01/201… Y          NA             NA              
#>  3 17-037243 2017-08     01/01/201… Y          NA             NA              
#>  4 17-027802 2017-08     01/01/201… Y          NA             NA              
#>  5 17-036999 2017-08     01/01/201… Y          NA             NA              
#>  6 17-000146 2017-08     01/01/201… Y          NA             NA              
#>  7 17-043830 2017-08     01/03/201… Y          NA             NA              
#>  8 17-042891 2017-08     01/04/201… Y          NA             NA              
#>  9 17-039404 2017-08     02/10/201… Y          NA             NA              
#> 10 17-012279 2017-08     02/14/201… Y          NA             NA              
#> 11 17-040862 2017-08     02/15/201… Y          NA             NA              
#> 12 17-040862 2017-08     02/15/201… NA         NA             NA              
#> 13 17-041667 2017-08     02/21/201… Y          NA             NA              
#> 14 17-008547 2017-08     02/21/201… Y          NA             NA              
#> 15 17-038102 2017-08     03/01/201… Y          NA             NA              
#> 16 17-043837 2017-08     03/15/201… Y          NA             NA              
#> 17 17-014252 2017-08     03/27/201… NA         NA             NA              
#> 18 17-039249 2017-08     03/31/201… Y          NA             NA              
#> 19 17-041006 2017-08     04/02/201… Y          NA             NA              
#> 20 17-041873 2017-08     04/12/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $December
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 18-000601 2017-12     01/01/199… Y          NA             NA              
#>  2 17-058893 2017-12     01/01/201… Y          NA             NA              
#>  3 17-059136 2017-12     01/01/201… Y          NA             NA              
#>  4 17-062864 2017-12     01/01/201… Y          NA             NA              
#>  5 17-058907 2017-12     01/01/201… Y          NA             NA              
#>  6 18-000313 2017-12     01/01/201… Y          NA             NA              
#>  7 17-059185 2017-12     01/23/201… Y          NA             NA              
#>  8 17-062934 2017-12     02/01/201… Y          NA             NA              
#>  9 17-063079 2017-12     02/17/201… Y          NA             NA              
#> 10 17-063078 2017-12     03/01/201… Y          NA             NA              
#> 11 17-062953 2017-12     03/13/201… Y          NA             NA              
#> 12 17-063247 2017-12     04/11/200… Y          NA             NA              
#> 13 17-023311 2017-12     04/12/201… NA         NA             NA              
#> 14 17-059618 2017-12     05/19/201… Y          NA             NA              
#> 15 18-000739 2017-12     05/22/201… Y          NA             NA              
#> 16 17-060377 2017-12     06/01/201… Y          NA             NA              
#> 17 17-059984 2017-12     06/01/201… Y          NA             NA              
#> 18 17-060136 2017-12     06/01/201… Y          NA             NA              
#> 19 17-059830 2017-12     06/12/201… Y          NA             NA              
#> 20 17-029925 2017-12     06/19/201… NA         NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $February
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-007856 2017-02     01/01/201… Y          NA             NA              
#>  2 17-009506 2017-02     01/01/201… Y          NA             NA              
#>  3 17-006283 2017-02     01/01/201… Y          NA             NA              
#>  4 17-008250 2017-02     01/01/201… Y          NA             NA              
#>  5 17-007727 2017-02     01/01/201… Y          NA             NA              
#>  6 17-008491 2017-02     01/01/201… Y          NA             NA              
#>  7 17-005479 2017-02     01/01/201… Y          NA             NA              
#>  8 17-006545 2017-02     01/01/201… Y          NA             NA              
#>  9 17-008706 2017-02     01/01/201… Y          NA             NA              
#> 10 17-008263 2017-02     01/01/201… Y          NA             NA              
#> 11 17-000225 2017-02     01/02/201… Y          NA             NA              
#> 12 17-009984 2017-02     01/02/201… Y          NA             NA              
#> 13 17-007276 2017-02     01/02/201… Y          NA             NA              
#> 14 17-008604 2017-02     01/02/201… Y          NA             NA              
#> 15 17-007601 2017-02     01/04/201… Y          NA             NA              
#> 16 17-005166 2017-02     01/04/201… Y          NA             NA              
#> 17 17-006240 2017-02     01/05/201… Y          NA             NA              
#> 18 17-008438 2017-02     01/05/201… Y          NA             NA              
#> 19 17-008601 2017-02     01/05/201… Y          NA             NA              
#> 20 17-007869 2017-02     01/06/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $January
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-005061 2017-01     01/01/200… Y          NA             NA              
#>  2 17-000381 2017-01     01/01/201… Y          NA             NA              
#>  3 17-001900 2017-01     01/01/201… Y          NA             NA              
#>  4 17-000294 2017-01     01/01/201… Y          NA             NA              
#>  5 17-002617 2017-01     01/01/201… Y          NA             NA              
#>  6 17-000104 2017-01     01/01/201… Y          NA             NA              
#>  7 17-000008 2017-01     01/01/201… Y          NA             NA              
#>  8 17-000055 2017-01     01/01/201… Y          NA             NA              
#>  9 17-000107 2017-01     01/01/201… Y          NA             NA              
#> 10 17-000071 2017-01     01/01/201… Y          NA             NA              
#> 11 17-000118 2017-01     01/01/201… Y          NA             NA              
#> 12 17-000009 2017-01     01/01/201… Y          NA             NA              
#> 13 17-000049 2017-01     01/01/201… Y          NA             NA              
#> 14 17-000010 2017-01     01/01/201… Y          NA             NA              
#> 15 17-000108 2017-01     01/01/201… Y          NA             NA              
#> 16 17-000014 2017-01     01/01/201… Y          NA             NA              
#> 17 17-000016 2017-01     01/01/201… Y          NA             NA              
#> 18 17-000065 2017-01     01/01/201… Y          NA             NA              
#> 19 17-000448 2017-01     01/01/201… Y          NA             NA              
#> 20 17-000012 2017-01     01/01/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $July
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-036556 2017-07     01/01/200… Y          NA             NA              
#>  2 17-036364 2017-07     01/01/201… Y          NA             NA              
#>  3 17-034387 2017-07     01/01/201… Y          NA             NA              
#>  4 17-037951 2017-07     01/01/201… Y          NA             NA              
#>  5 17-033635 2017-07     01/07/201… Y          NA             NA              
#>  6 17-035662 2017-07     01/08/201… Y          NA             NA              
#>  7 17-034149 2017-07     01/09/201… Y          NA             NA              
#>  8 17-037405 2017-07     01/11/201… Y          NA             NA              
#>  9 17-033639 2017-07     01/15/201… Y          NA             NA              
#> 10 17-033638 2017-07     01/25/201… Y          NA             NA              
#> 11 17-033637 2017-07     01/29/201… Y          NA             NA              
#> 12 17-035879 2017-07     01/30/201… Y          NA             NA              
#> 13 17-036825 2017-07     02/07/201… Y          NA             NA              
#> 14 17-033972 2017-07     02/07/201… Y          NA             NA              
#> 15 17-006979 2017-07     02/12/201… Y          NA             NA              
#> 16 17-008993 2017-07     02/24/201… Y          NA             NA              
#> 17 17-034199 2017-07     03/02/201… Y          NA             NA              
#> 18 17-010983 2017-07     03/07/201… Y          NA             NA              
#> 19 17-010783 2017-07     03/07/201… Y          NA             NA              
#> 20 17-035199 2017-07     03/09/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $June
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-024541 2017-06     01/01/200… Y          NA             NA              
#>  2 17-031801 2017-06     01/01/200… Y          NA             NA              
#>  3 17-028791 2017-06     01/01/201… Y          NA             NA              
#>  4 17-031759 2017-06     01/01/201… Y          NA             NA              
#>  5 17-028681 2017-06     01/01/201… Y          NA             NA              
#>  6 17-017054 2017-06     01/01/201… NA         NA             NA              
#>  7 17-025159 2017-06     01/01/201… Y          NA             NA              
#>  8 17-031573 2017-06     01/01/201… Y          NA             NA              
#>  9 17-031193 2017-06     01/01/201… Y          NA             NA              
#> 10 17-001442 2017-06     01/03/201… Y          NA             NA              
#> 11 17-028963 2017-06     01/15/201… Y          NA             NA              
#> 12 17-026459 2017-06     01/30/201… Y          NA             NA              
#> 13 17-030149 2017-06     02/02/201… Y          NA             NA              
#> 14 17-032447 2017-06     02/09/201… Y          NA             NA              
#> 15 17-031620 2017-06     02/11/201… Y          NA             NA              
#> 16 17-007863 2017-06     02/18/201… Y          NA             NA              
#> 17 17-007863 2017-06     02/18/201… Y          NA             NA              
#> 18 17-007863 2017-06     02/18/201… Y          NA             NA              
#> 19 17-007863 2017-06     02/18/201… Y          NA             NA              
#> 20 17-007863 2017-06     02/18/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $March
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-008256 2017-03     01/01/198… Y          NA             NA              
#>  2 17-007427 2017-03     01/01/200… Y          NA             NA              
#>  3 17-014231 2017-03     01/01/200… Y          NA             NA              
#>  4 17-012773 2017-03     01/01/200… Y          NA             NA              
#>  5 17-013173 2017-03     01/01/201… Y          NA             NA              
#>  6 17-010723 2017-03     01/01/201… Y          NA             NA              
#>  7 17-013076 2017-03     01/01/201… Y          NA             NA              
#>  8 17-013656 2017-03     01/01/201… Y          NA             NA              
#>  9 17-013667 2017-03     01/01/201… Y          NA             NA              
#> 10 17-010239 2017-03     01/01/201… Y          NA             NA              
#> 11 17-012990 2017-03     01/01/201… Y          NA             NA              
#> 12 17-010740 2017-03     01/01/201… Y          NA             NA              
#> 13 17-012770 2017-03     01/01/201… Y          NA             NA              
#> 14 17-014168 2017-03     01/03/201… Y          NA             NA              
#> 15 17-014754 2017-03     01/03/201… Y          NA             NA              
#> 16 17-014756 2017-03     01/03/201… Y          NA             NA              
#> 17 17-011263 2017-03     01/03/201… Y          NA             NA              
#> 18 17-014179 2017-03     01/05/201… Y          NA             NA              
#> 19 17-000913 2017-03     01/06/201… Y          NA             NA              
#> 20 17-014753 2017-03     01/07/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $May
#> # A tibble: 20 x 26
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-022207 2017-05     5/10/2017… Y          NA             NA              
#>  2 17-021969 2017-05     5/9/2017 … Y          NA             NA              
#>  3 17-022476 2017-05     5/12/2017… Y          NA             NA              
#>  4 17-022597 2017-05     5/12/2017… Y          NA             NA              
#>  5 17-022571 2017-05     5/12/2017… Y          NA             NA              
#>  6 17-022648 2017-05     5/12/2017… Y          NA             NA              
#>  7 17-022455 2017-05     5/11/2017… Y          NA             NA              
#>  8 17-022598 2017-05     5/12/2017… Y          NA             NA              
#>  9 17-022695 2017-05     5/13/2017… Y          NA             NA              
#> 10 17-021904 2017-05     5/8/2017 … Y          NA             NA              
#> 11 17-022783 2017-05     5/13/2017… Y          NA             NA              
#> 12 17-022330 2017-05     4/25/2017… Y          NA             NA              
#> 13 17-022126 2017-05     5/10/2017… Y          NA             NA              
#> 14 17-022723 2017-05     5/9/2017 … Y          NA             NA              
#> 15 17-022687 2017-05     5/13/2017… Y          NA             NA              
#> 16 17-022021 2017-05     5/9/2017 … Y          NA             NA              
#> 17 17-015885 2017-05     4/1/2016 … Y          NA             NA              
#> 18 17-022143 2017-05     5/8/2017 … Y          NA             NA              
#> 19 17-022196 2017-05     5/10/2017… Y          NA             NA              
#> 20 17-022900 2017-05     5/14/2017… Y          NA             NA              
#> # … with 20 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   i_leads_add <chr>, i_leads_approve <chr>, district <chr>, beat <chr>,
#> #   description <chr>, ileads_address <chr>, ileads_street <chr>,
#> #   neighborhood <chr>, i_leads_asg <chr>, i_leads_type <chr>,
#> #   location_name <chr>, location_comment <chr>, cad_address <chr>,
#> #   cad_street <chr>, x_coord <chr>, y_coord <chr>, date_crime_coded <chr>
#> 
#> $November
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-057965 2017-11     01/01/201… Y          NA             NA              
#>  2 17-001065 2017-11     01/06/201… NA         NA             NA              
#>  3 17-001065 2017-11     01/06/201… NA         NA             NA              
#>  4 17-004292 2017-11     01/26/201… NA         NA             NA              
#>  5 16-004824 2017-11     01/30/201… NA         NA             NA              
#>  6 17-005402 2017-11     02/02/201… NA         NA             NA              
#>  7 17-057568 2017-11     02/15/201… Y          NA             NA              
#>  8 17-008198 2017-11     02/20/201… NA         NA             NA              
#>  9 17-058452 2017-11     02/26/200… Y          NA             NA              
#> 10 17-054117 2017-11     03/01/201… Y          NA             NA              
#> 11 17-054676 2017-11     03/01/201… Y          NA             NA              
#> 12 17-047698 2017-11     03/03/201… Y          NA             NA              
#> 13 17-011024 2017-11     03/08/201… NA         NA             NA              
#> 14 17-054639 2017-11     03/30/201… Y          NA             NA              
#> 15 17-055342 2017-11     04/01/201… Y          NA             NA              
#> 16 17-058296 2017-11     04/01/201… Y          NA             NA              
#> 17 17-057820 2017-11     04/05/201… Y          NA             NA              
#> 18 17-024948 2017-11     04/17/201… Y          NA             NA              
#> 19 17-019083 2017-11     04/23/201… NA         NA             NA              
#> 20 17-022195 2017-11     05/10/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $October
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-049033 2017-10     01/01/200… Y          NA             NA              
#>  2 17-050757 2017-10     01/01/201… Y          NA             NA              
#>  3 17-052030 2017-10     01/01/201… Y          NA             NA              
#>  4 17-052030 2017-10     01/01/201… Y          NA             NA              
#>  5 17-049705 2017-10     01/01/201… Y          NA             NA              
#>  6 17-001344 2017-10     01/09/201… Y          NA             NA              
#>  7 17-001344 2017-10     01/09/201… NA         NA             NA              
#>  8 17-001344 2017-10     01/09/201… NA         NA             NA              
#>  9 17-051540 2017-10     01/15/201… Y          NA             NA              
#> 10 17-049595 2017-10     02/03/201… Y          NA             NA              
#> 11 17-006678 2017-10     02/10/201… NA         NA             NA              
#> 12 17-048535 2017-10     04/04/201… Y          NA             NA              
#> 13 17-048535 2017-10     04/04/201… Y          NA             NA              
#> 14 17-017618 2017-10     04/14/201… Y          NA             NA              
#> 15 17-017618 2017-10     04/14/201… Y          NA             NA              
#> 16 17-017618 2017-10     04/14/201… Y          NA             NA              
#> 17 17-051182 2017-10     04/19/201… Y          NA             NA              
#> 18 17-049901 2017-10     05/02/201… Y          NA             NA              
#> 19 17-049903 2017-10     05/03/201… Y          NA             NA              
#> 20 17-052814 2017-10     05/04/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> $September
#> # A tibble: 20 x 20
#>    complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>    <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#>  1 17-044036 2017-09     01/01/200… Y          NA             NA              
#>  2 17-044253 2017-09     01/01/201… Y          NA             NA              
#>  3 17-043951 2017-09     01/10/201… Y          NA             NA              
#>  4 17-048470 2017-09     01/17/201… Y          NA             NA              
#>  5 17-043154 2017-09     02/18/200… Y          NA             NA              
#>  6 17-040776 2017-09     03/01/201… Y          NA             NA              
#>  7 17-046259 2017-09     03/18/201… Y          NA             NA              
#>  8 17-046261 2017-09     03/19/201… Y          NA             NA              
#>  9 17-045409 2017-09     03/20/201… Y          NA             NA              
#> 10 17-015017 2017-09     03/31/201… Y          NA             NA              
#> 11 17-015017 2017-09     03/31/201… Y          NA             NA              
#> 12 17-015017 2017-09     03/31/201… Y          NA             NA              
#> 13 17-015017 2017-09     03/31/201… Y          NA             NA              
#> 14 17-015017 2017-09     03/31/201… NA         NA             NA              
#> 15 17-044925 2017-09     04/01/201… Y          NA             NA              
#> 16 17-045308 2017-09     04/05/201… Y          NA             NA              
#> 17 17-047579 2017-09     04/14/201… Y          NA             NA              
#> 18 14-015756 2017-09     04/15/201… NA         NA             NA              
#> 19 17-045407 2017-09     04/15/201… Y          NA             NA              
#> 20 17-017939 2017-09     04/16/201… Y          NA             NA              
#> # … with 14 more variables: count <chr>, flag_cleanup <chr>, crime <chr>,
#> #   district <chr>, description <chr>, ileads_address <chr>,
#> #   ileads_street <chr>, neighborhood <chr>, location_name <chr>,
#> #   location_comment <chr>, cad_address <chr>, cad_street <chr>, x_coord <chr>,
#> #   y_coord <chr>
#> 
#> attr(,"class")
#> [1] "list"         "cs_year_list"