Identify Missing Coordinates

cs_missingXY compares X and Y coordinates and adds a logical column that identifies observations that are missing coordinate data.

cs_missingXY(.data, varX, varY, newVar)

Arguments

.data
A tibble or data frame

varX
Name of column containing x coordinate data

varY
Name of column containing y coordinate data

newVar
Name of new column that is TRUE if coordinate data are missing and FALSE otherwise.

Value

A tibble or data frame with a logical vector appended to it.

Examples

cs_missingXY(january2018, varX = XCoord, varY = YCoord, newVar = missingXY)

#> # A tibble: 3,825 x 21
#>    Complaint CodedMonth DateOccur FlagCrime FlagUnfounded FlagAdministrat… Count
#>    <chr>     <chr>      <chr>     <chr>     <lgl>         <lgl>            <dbl>
#>  1 18-001804 2018-01    01/01/20… Y         NA            NA                   1
#>  2 18-002209 2018-01    01/01/20… Y         NA            NA                   1
#>  3 18-000063 2018-01    01/01/20… Y         NA            NA                   1
#>  4 18-000018 2018-01    01/01/20… Y         NA            NA                   1
#>  5 18-000003 2018-01    01/01/20… Y         NA            NA                   1
#>  6 18-000132 2018-01    01/01/20… Y         NA            NA                   1
#>  7 18-000010 2018-01    01/01/20… Y         NA            NA                   1
#>  8 18-000010 2018-01    01/01/20… Y         NA            NA                   1
#>  9 18-003411 2018-01    01/01/20… Y         NA            NA                   1
#> 10 18-005259 2018-01    01/01/20… Y         NA            NA                   1
#> # … with 3,815 more rows, and 14 more variables: FlagCleanup <lgl>,
#> #   Crime <dbl>, District <dbl>, Description <chr>, ILEADSAddress <dbl>,
#> #   ILEADSStreet <chr>, Neighborhood <dbl>, LocationName <chr>,
#> #   LocationComment <chr>, CADAddress <dbl>, CADStreet <chr>, XCoord <dbl>,
#> #   YCoord <dbl>, missingXY <chr>