Crimes in St. Louis, January 2018

A data set containing all reported crimes in St. Louis, Missouri during January 2018.

data(january2018)

Format

A tibble with 3825 rows and 20 variables:

complaint

complaint record number

coded_month

year and month crime reported

date_occur

date and time of crime

flag_crime

Returns a Y when a crime occurred

flag_unfounded

Reported crime was investigated and determined to be unfounded

flag_administrative

Reported crime had a change in classification

count

Returns a 1 or -1 for counting purposes

flag_cleanup

Returns information if administrative cleanup occured

crime

Uniform Crime Reporting code

district

Number corresponding to the police determined district

description

Name of the crime

ileads_address

I/Leads system address

ileads_street

I/Leads system street

neighborhood

Number corresponding to a neighborhood

location_name

Common “Location Name” (i.e. Zoo, Scottrade Center, etc.)

location_comment

Information to provide context to the location (i.e. Alley, Restaraunt Name)

cad_address

The Computer-Aided Dispatch address is the reported address by the 911 caller

cad_street

The Computer-Aided Dispatch street is the reported street by the 911 caller

x_coord

X-coordinates in the NAD83 format

y_coord

Y-coordinates in the NAD83 format

Source

St. Louis Metropolitan Police Department

Examples

str(january2018)

#> tibble [3,825 × 20] (S3: tbl_df/tbl/data.frame)
#>  $ complaint          : chr [1:3825] "18-001804" "18-002209" "18-000063" "18-000018" ...
#>  $ coded_month        : chr [1:3825] "2018-01" "2018-01" "2018-01" "2018-01" ...
#>  $ date_occur         : chr [1:3825] "01/01/2018 00:01" "01/01/2018 00:01" "01/01/2018 00:01" "01/01/2018 00:01" ...
#>  $ flag_crime         : chr [1:3825] "Y" "Y" "Y" "Y" ...
#>  $ flag_unfounded     : chr [1:3825] NA NA NA NA ...
#>  $ flag_administrative: chr [1:3825] NA NA NA NA ...
#>  $ count              : int [1:3825] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ flag_cleanup       : chr [1:3825] NA NA NA NA ...
#>  $ crime              : int [1:3825] 31111 142320 67601 41011 41011 142320 41021 41011 142320 67701 ...
#>  $ district           : int [1:3825] 1 5 4 2 3 6 5 5 4 1 ...
#>  $ description        : chr [1:3825] "ROBBERY-HIGHWAY  /FIREARM USED/SUCCESSFUL" "DESTRUCTION OF PROPERTY-MALICIOUS/PRIV PROP" "LARCENY-FROM BUILDING $500 - $24,999" "AGG.ASSAULT-FIREARM/CITIZEN ADULT 1ST DEGREE" ...
#>  $ ileads_address     : num [1:3825] 6431 5506 2543 4000 3149 ...
#>  $ ileads_street      : chr [1:3825] "IDAHO AVE" "ETZEL AVE" "N GRAND BLVD" "DE TONTY ST" ...
#>  $ neighborhood       : int [1:3825] 1 48 59 27 25 71 38 38 36 5 ...
#>  $ location_name      : chr [1:3825] NA NA "THE BOULEVARD" NA ...
#>  $ location_comment   : chr [1:3825] "ON STREET IN FRONT OF 6431 IDAHO" NA NA NA ...
#>  $ cad_address        : num [1:3825] 6431 5506 2543 4000 3149 ...
#>  $ cad_street         : chr [1:3825] "IDAHO" "ETZEL" "GRAND" "DE TONTY" ...
#>  $ x_coord            : num [1:3825] 890047 884088 899362 892323 896040 ...
#>  $ y_coord            : num [1:3825] 992586 1030426 1026965 1013550 1008823 ...

head(january2018)

#> # A tibble: 6 x 20
#>   complaint coded_month date_occur flag_crime flag_unfounded flag_administra…
#>   <chr>     <chr>       <chr>      <chr>      <chr>          <chr>           
#> 1 18-001804 2018-01     01/01/201… Y          NA             NA              
#> 2 18-002209 2018-01     01/01/201… Y          NA             NA              
#> 3 18-000063 2018-01     01/01/201… Y          NA             NA              
#> 4 18-000018 2018-01     01/01/201… Y          NA             NA              
#> 5 18-000003 2018-01     01/01/201… Y          NA             NA              
#> 6 18-000132 2018-01     01/01/201… Y          NA             NA              
#> # … with 14 more variables: count <int>, flag_cleanup <chr>, crime <int>,
#> #   district <int>, description <chr>, ileads_address <dbl>,
#> #   ileads_street <chr>, neighborhood <int>, location_name <chr>,
#> #   location_comment <chr>, cad_address <dbl>, cad_street <chr>, x_coord <dbl>,
#> #   y_coord <dbl>