Filter Crimes

cs_filter_crime can be used to subset based on specific single UCR categories or common groupings. This can be used on any police department's data where codes like 31111 (robbery with a firearm) or 142320 (malicious destruction of property) are used to identify crimes.

cs_filter_crime(.data, var, crime)

Arguments

.data
A tibble or data frame

var
Name of variable with 5 or 6 digit crime codes

crime
A string describing the crime type to be identified

Value

A subset object with only the specified crimes

Details

The categories used here are derived from the U.S. Federal Bureau of Investigation's Uniform Crime Reporting codes. Valid inputs for the crime argument are as follows:

"violent"

Violent crimes (homicide, rape, aggrevated assault, and robbery)

"property"

Property crimes (burglary, larceny, larceny of a motor vehicle, and arson)

"part 1"

All violent and property crimes

"homicide"

"murder" is also acceptible as input as is UCR code 1

"rape"

"forcible rape" is also acceptible as input as is UCR code 2

"robbery"

UCR code 3 is also acceptible input

"agg assualt"

"aggrevated assualt" is also acceptible as input as is UCR code 4

"burglary"

UCR code 5 is also acceptible input

"larceny-theft"

"larceny" and "theft" are also acceptible inputs as is UCR code 6

"mv theft"

"motor vehicle theft", "motor vehicle larceny", and "mv larceny" are also acceptible inputs as input as is UCR code 7

"arson"

UCR code 8 is also acceptible input

"part 2"

All other crimes

"assault"

"other assaults" is also acceptible input as is UCR code 9

"forgery"

"forgery and counterfeiting" is also acceptible input as is UCR code 10

"fraud"

UCR code 11 is also acceptible input

"embezzlement"

UCR code 12 is also acceptible input

"stolen prop"

"stolen property" is also acceptible input as is UCR code 13

"vandalism"

UCR code 14 is also acceptible input

"weapons"

UCR code 15 is also acceptible input

"prostitution"

"prostitution and commercialized vice" is also acceptible input as is UCR code 16

"sex offenses"

UCR code 17 is also acceptible input

"drugs"

"drug abuse violations" is also acceptible input as is UCR code 18

"gambling"

UCR code 19 is also acceptible input

"family"

"offenses against the family and children" is also acceptible input as is UCR code 20

"dwi"

"driving under the influence" is also acceptible input as is UCR code 21

"liquor laws"

UCR code 22 is also acceptible input

"drunkenness"

UCR code 23 is also acceptible input

"discon"

"disorderly conduct" is also acceptible input as is UCR code 24

"vagrancy"

UCR code 25 is also acceptible input

"other"

"all other offenses" is also acceptible input as is UCR code 26

"suspicion"

UCR code 27 is also acceptible input

"curfew"

"curfew and loitering laws-persons under 18" is also acceptible input as is UCR code 28

"runaway"

"runaways-persons under 18" is also acceptible input as is UCR code 29

Examples

# load example data
testData <- january2018

# subset data to retain only violent crimes
testData <- cs_filter_crime(testData, var = Crime, crime = "violent")