Crimes in St. Louis, January 2018

A data set containing all reported crimes in St. Louis, Missouri during January 2018.

data(january2018)

Format

A tibble with 3825 rows and 20 variables:

Complaint

complaint record number

CodedMonth

year and month crime reported

DateOccur

date and time of crime

FlagCrime

Returns a Y when a crime occured

FlagUnfounded

Reported crime was investigated and determined to be unfounded

FlagAdministrative

Reported crime had a change in classification

Count

Returns a 1 or -1 for counting purposes

FlagCleanup

Returns information if administrative cleanup occured

Crime

Uniform Crime Reporting code

District

Number corresponding to the police determined district

Description

Name of the crime

ILEADSAddress

I/Leads system address

ILEADSStreet

I/Leads system street

Neighborhood

Number corresponding to a neighborhood

LocationName

Common “Location Name” (i.e. Zoo, Scottrade Center, etc.)

LocationComment

Information to provide context to the location (i.e. Alley, Restaraunt Name)

CADAddress

The Computer-Aided Dispatch address is the reported address by the 911 caller

CADStreet

The Computer-Aided Dispatch street is the reported street by the 911 caller

XCoord

X-coordinates in the NAD83 format

YCoord

Y-coordinates in the NAD83 format

Source

St. Louis Metropolitan Police Department

Examples

str(january2018)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	3825 obs. of  20 variables:
#>  $ Complaint         : chr  "18-001804" "18-002209" "18-000063" "18-000018" ...
#>  $ CodedMonth        : chr  "2018-01" "2018-01" "2018-01" "2018-01" ...
#>  $ DateOccur         : chr  "01/01/2018 00:01" "01/01/2018 00:01" "01/01/2018 00:01" "01/01/2018 00:01" ...
#>  $ FlagCrime         : chr  "Y" "Y" "Y" "Y" ...
#>  $ FlagUnfounded     : logi  NA NA NA NA NA NA ...
#>  $ FlagAdministrative: logi  NA NA NA NA NA NA ...
#>  $ Count             : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ FlagCleanup       : logi  NA NA NA NA NA NA ...
#>  $ Crime             : num  31111 142320 67601 41011 41011 ...
#>  $ District          : num  1 5 4 2 3 6 5 5 4 1 ...
#>  $ Description       : chr  "ROBBERY-HIGHWAY  /FIREARM USED/SUCCESSFUL" "DESTRUCTION OF PROPERTY-MALICIOUS/PRIV PROP" "LARCENY-FROM BUILDING $500 - $24,999" "AGG.ASSAULT-FIREARM/CITIZEN ADULT 1ST DEGREE" ...
#>  $ ILEADSAddress     : num  6431 5506 2543 4000 3149 ...
#>  $ ILEADSStreet      : chr  "IDAHO AVE" "ETZEL AVE" "N GRAND BLVD" "DE TONTY ST" ...
#>  $ Neighborhood      : num  1 48 59 27 25 71 38 38 36 5 ...
#>  $ LocationName      : chr  NA NA "THE BOULEVARD" NA ...
#>  $ LocationComment   : chr  "ON STREET IN FRONT OF 6431 IDAHO" NA NA NA ...
#>  $ CADAddress        : num  6431 5506 2543 4000 3149 ...
#>  $ CADStreet         : chr  "IDAHO" "ETZEL" "GRAND" "DE TONTY" ...
#>  $ XCoord            : num  890047 884088 899362 892323 896040 ...
#>  $ YCoord            : num  992586 1030426 1026965 1013550 1008823 ...

head(january2018)

#> # A tibble: 6 x 20
#>   Complaint CodedMonth DateOccur FlagCrime FlagUnfounded FlagAdministrat… Count
#>   <chr>     <chr>      <chr>     <chr>     <lgl>         <lgl>            <dbl>
#> 1 18-001804 2018-01    01/01/20… Y         NA            NA                   1
#> 2 18-002209 2018-01    01/01/20… Y         NA            NA                   1
#> 3 18-000063 2018-01    01/01/20… Y         NA            NA                   1
#> 4 18-000018 2018-01    01/01/20… Y         NA            NA                   1
#> 5 18-000003 2018-01    01/01/20… Y         NA            NA                   1
#> 6 18-000132 2018-01    01/01/20… Y         NA            NA                   1
#> # … with 13 more variables: FlagCleanup <lgl>, Crime <dbl>, District <dbl>,
#> #   Description <chr>, ILEADSAddress <dbl>, ILEADSStreet <chr>,
#> #   Neighborhood <dbl>, LocationName <chr>, LocationComment <chr>,
#> #   CADAddress <dbl>, CADStreet <chr>, XCoord <dbl>, YCoord <dbl>